The new parser that does also constant conversion.
We use the transformer ParsecT directly, wrapped around a suitable
state monad whose state is a pair consisting of a scalar environment
that is augmented every time a constant is converted, and a count that
is used to generate unique variable names and that is incremented
every time a new name is generated.

The module VL.Convert is not necessary anymore.