Re-factored lifting of dyadic operations.
First, I thought I could remove some code duplication with type
classes.  I introduced a class, which I called Abstractable, with
three methods: `convert', allowing to view concrete values as abstract
values, `extract', for extraction of concrete values from abstract
values, and `abstract', which allows us to generalize a concrete value
to an abstract value, e.g., True to AbstractBoolean.  I then defined
instances of Abstractable for Float and Bool and tried to define a
generic lifting function `liftOp'.  Alas, I couldn't get this to work,
so I decided to instead associate with any type a dictionary of
functions as above and to pass this dictionary explicitly to `liftOp'.