In order to marry recursion and `real', AbstractTop must be replaced with AbstractBottom.
`unifyValues' has been given a better name `joinValues' because it
implements the join in the join-semilattice of abstract values.  The
join doesn't always exist!  As a consequence, expressions like (if (<
x y) #f (- y x)) cannot be analyzed.  It can be argued that such
expressions should not be allowed at all, as they are ill-typed (the
branches of the conditional have different types); alternatively, we
can say that such expressions are not union-free and our analysis is
only capable of analyzing union-free programs.