The code generator is almost complete; alas, it doesn't work, even on pretty simple examples.
It is not clear what to do with void values.  E.g., in

((lambda (x) x) (real 3))

the closure is void (because its environment is empty), however, it
yields non-void results when applied to the abstract real, so some
code must be generated for it.  But it can't be!  Because the code
would include a function that takes two arguments, a closure and a
double, but there is no C type associated with the closure because
it's void.