Added 'IfProc', but it doesn't work yet.
The problem is that I force thunks differently (not by applying them
to ()), so for each thunk I have to introduce a function that takes
that thunk as an argument and forces it (its body is the body of the
thunk compiled in the environment of the thunk).  Which means I need
one more lookup table (or, probably, I could just name these functions
F (thunk, AbstractScalar Nil), haven't decided yet what would be
cleaner; probably it makes sense to have a different set of names).