Prepare.hs and Uniquify.hs are now called Read.hs and Rename.hs, respectively.
Rename.hs exports a function 'alphaRename' that renames also free
variables in an expression (in contrast to 'uniquify' that was
renaming only bound variables).  The code generator requires that
because there are name clashes between VL and C functions.

Read.hs exports a function 'read' that does a little more than
'prepare' used to do.  Namely, it adds the primitives to the
environment, and renames the variables in the environment
to match the names of free variables in the expression.