Rewrote the code generator in monadic style.
Instead of precomputing and threading through the lookup tables, they
are made part of the code generator state, and thus are carried around
and are computed on demand.  Also, IF is treated differently; better,
but still not entirely correct.  The problem is that thunks are forced
not by applying to (), but by evaluating the body of the thunk in the
environment of the thunk (the difference is essential for the flow
analysis).  Therefore, there are closures that are not refined as such
during the flow analysis.  So they must be treated specially during
compilation.  I don't yet fully understand how.