Added LETREC.
Surprisingly, the same pattern as in both the concrete and abstract
evaluator works: 'compileExpr' of LETREC simply pushes LETREC inwards
and recursively call 'compileExpr' on the obtained expression.  What I
did have to change is enumeration of closure-value pairs.  At the end
of the day it is very simple: take all pairs (abstract closure,
abstract value) for which 'refineApply' returns non-bottom.